stages/source/requirements.txt
kafka-python>=2,<3